+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Updated datasource ID instead of name and added support variables (#229)
 - Updated sticky header for Scenes dashboards (#230)
+- Updated display favorites icon (#234)
+- Added not supports variable type Alert message (#235)
 
 ## 3.6.0 (2024-10-24)